+++ plugins/@sanity/personalization-plugin/tsdown.config.ts
@@ -4,5 +4,4 @@ import type {UserConfig} from 'tsdown'
 export default defineConfig({
   entry: ['./src/index.ts', './src/launchDarkly/index.ts', './src/growthbook/index.ts'],
   reactCompiler: true,
-  dts: {tsgo: true},
 }) satisfies Promise<UserConfig>