+++ plugins/sanity-plugin-mux-input/tsdown.config.ts
@@ -7,6 +7,5 @@ export default defineConfig({
   entry: ['./src/_exports/index.ts'],
   styledComponents: true,
   reactCompiler: true,
-  dts: {tsgo: true},
   define: {__PKG_VERSION__: JSON.stringify(pkg.version)},
 }) satisfies Promise<UserConfig>