+++ plugins/@sanity/google-maps-input/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sanity/google-maps-input
 
+## 6.1.4
+
+### Patch Changes
+
+- [#1571](https://github.com/sanity-io/plugins/pull/1571) [`52975b2`](https://github.com/sanity-io/plugins/commit/52975b2f0d4ea5086c800b2ce16190b862284a95) Thanks [@stipsan](https://github.com/stipsan)! - The vanilla-extract CSS bundle is now built by `@sanity/vanilla-extract-tsdown-plugin` (via `@sanity/tsdown-config` 0.14.0) instead of `@vanilla-extract/rollup-plugin`. `dist/bundle.css` is slightly smaller, the `dist/bundle.css.map` sourcemap is no longer emitted (aligned with `@tsdown/css` behavior), and the `bundle.css.js` node shim is now a comment-only no-op module instead of `export default ""`
+
+- [#1571](https://github.com/sanity-io/plugins/pull/1571) [`52975b2`](https://github.com/sanity-io/plugins/commit/52975b2f0d4ea5086c800b2ce16190b862284a95) Thanks [@stipsan](https://github.com/stipsan)! - fix(deps): update tsdown to ^0.22.7 and @sanity/tsdown-config to ^0.14.0
+
 ## 6.1.3
 
 ### Patch Changes